+++ ui/src/AccessRevoke.tsx
@@ -6,12 +6,13 @@ import { getNoTTLVRequest, postNoTTLVRequest } from "./utils";
 interface AccessRevokeFormData {
     user_id: string;
     unique_identifier: string;
-    operation_types: Array<"create" | "get" | "encrypt" | "decrypt" | "import" | "revoke" | "locate" | "rekey" | "destroy">;
+    operation_types: Array<"create" | "get" | "getattributes" | "encrypt" | "decrypt" | "import" | "revoke" | "locate" | "rekey" | "destroy">;
     revoke_create_access_right: boolean;
 }
 
 const KMIP_OPERATIONS = [
     { label: "Get", value: "get" },
+    { label: "GetAttributes", value: "getattributes" },
     { label: "Encrypt", value: "encrypt" },
     { label: "Decrypt", value: "decrypt" },
     { label: "Revoke", value: "revoke" },